+++ isitfit/cost/ec2_optimize.py
@@ -225,7 +225,7 @@ class ReporterOptimizeEc2(ReporterBase):
 
   def __init__(self):
     # for final csv file
-    self.csv_fn_final = None
+    # DEPRECATED # self.csv_fn_final = None
 
     # members that will contain the results of the optimization
     self.df_sort = None
@@ -239,12 +239,12 @@ class ReporterOptimizeEc2(ReporterBase):
 
     # process
     self._after_all()
-    self._storecsv_all()
+    # DEPRECATED # self._storecsv_all()
 
     # save to context for aggregator
     context_all['df_sort'] = self.df_sort
     context_all['sum_val'] = self.sum_val
-    context_all['csv_fn_final'] = self.csv_fn_final
+    # DEPRECATED # context_all['csv_fn_final'] = self.csv_fn_final
 
     # done
     return context_all
@@ -300,70 +300,72 @@ class ReporterOptimizeEc2(ReporterBase):
     self.sum_val = df_all.savings.sum()
 
 
-  def _storecsv_all(self, *args, **kwargs):
-      if self.df_sort is None:
-        return
+# DEPRECATED
+#  def _storecsv_all(self, *args, **kwargs):
+#      if self.df_sort is None:
+#        return
+#
+#      import tempfile
+#      from isitfit.dotMan import DotMan
+#      with tempfile.NamedTemporaryFile(prefix='isitfit-full-ec2-', suffix='.csv', delete=False, dir=DotMan().tempdir()) as  csv_fh_final:
+#        self.csv_fn_final = csv_fh_final.name
+#        logger.debug(colored("Saving final results to %s"%csv_fh_final.name, "cyan"))
+#        self.df_sort.to_csv(csv_fh_final.name, index=False)
+#        logger.debug(colored("Save complete", "cyan"))
 
-      import tempfile
-      from isitfit.dotMan import DotMan
-      with tempfile.NamedTemporaryFile(prefix='isitfit-full-ec2-', suffix='.csv', delete=False, dir=DotMan().tempdir()) as  csv_fh_final:
-        self.csv_fn_final = csv_fh_final.name
-        logger.debug(colored("Saving final results to %s"%csv_fh_final.name, "cyan"))
-        self.df_sort.to_csv(csv_fh_final.name, index=False)
-        logger.debug(colored("Save complete", "cyan"))
 
-
-  def display(self, context_all):
-    if self.df_sort is None:
-      logger.info(colored("No EC2 instances found", "red"))
-      return context_all
-
-    # display
-    # Edit 2019-09-25 just show the full list. Will add filtering later. This way it's less ambiguous when all instances are "Normal"
-    # self.df_sort.dropna(subset=['recommended_type'], inplace=True)
-    
-    # if no recommendations
-    if self.df_sort.shape[0]==0:
-      logger.info(colored("No optimizations from isitfit for this AWS EC2 account", "red"))
-      return context_all
-    
-    # if there are recommendations, show them
-    sum_comment = "extra cost" if self.sum_val>0 else "savings"
-    sum_color = "red" if self.sum_val>0 else "green"
-
-    import click
-    #logger.info("Optimization based on the following CPU thresholds:")
-    #logger.info(self.thresholds)
-    #logger.info("")
-    click.echo(colored("Recommended %s: %0.0f $ (over the next 3 months)"%(sum_comment, self.sum_val), sum_color))
-    click.echo("")
-
-    # display dataframe
-    from isitfit.utils import display_df
-    display_df(
-      "Recommended EC2 size changes",
-      self.df_sort,
-      self.csv_fn_final,
-      self.df_sort.shape,
-      logger
-    )
-
-#    with pd.option_context("display.max_columns", 10):
-#      logger.info("Details")
-#      if self.df_sort.shape[0]<=10:
-#        logger.info(df2tabulate(self.df_sort))
-#      else:
-#        logger.info(df2tabulate(self.df_sort.head(n=5)))
-#        logger.info("...")
-#        logger.info(df2tabulate(self.df_sort.tail(n=5)))
-#        logger.info("")
-#        logger.info(colored("Table originally with %i rows is truncated for top and bottom 5 only."%self.df_sort.shape[0], "cyan"))
-#        logger.info(colored("Consider filtering it with --n=x for the 1st x results or --filter-tags=foo using a value from your own EC2 tags.", "cyan"))
-
-    if self.analyzer.n!=-1:
-      logger.info(colored("This table has been filtered for only the 1st %i underused results"%self.analyzer.n, "cyan"))
-
-    return context_all
+# DEPRECATED
+#  def display(self, context_all):
+#    if self.df_sort is None:
+#      logger.info(colored("No EC2 instances found", "red"))
+#      return context_all
+#
+#    # display
+#    # Edit 2019-09-25 just show the full list. Will add filtering later. This way it's less ambiguous when all instances are "Normal"
+#    # self.df_sort.dropna(subset=['recommended_type'], inplace=True)
+#    
+#    # if no recommendations
+#    if self.df_sort.shape[0]==0:
+#      logger.info(colored("No optimizations from isitfit for this AWS EC2 account", "red"))
+#      return context_all
+#    
+#    # if there are recommendations, show them
+#    sum_comment = "extra cost" if self.sum_val>0 else "savings"
+#    sum_color = "red" if self.sum_val>0 else "green"
+#
+#    import click
+#    #logger.info("Optimization based on the following CPU thresholds:")
+#    #logger.info(self.thresholds)
+#    #logger.info("")
+#    click.echo(colored("Recommended %s: %0.0f $ (over the next 3 months)"%(sum_comment, self.sum_val), sum_color))
+#    click.echo("")
+#
+#    # display dataframe
+#    from isitfit.utils import display_df
+#    display_df(
+#      "Recommended EC2 size changes",
+#      self.df_sort,
+#      self.csv_fn_final,
+#      self.df_sort.shape,
+#      logger
+#    )
+#
+##    with pd.option_context("display.max_columns", 10):
+##      logger.info("Details")
+##      if self.df_sort.shape[0]<=10:
+##        logger.info(df2tabulate(self.df_sort))
+##      else:
+##        logger.info(df2tabulate(self.df_sort.head(n=5)))
+##        logger.info("...")
+##        logger.info(df2tabulate(self.df_sort.tail(n=5)))
+##        logger.info("")
+##        logger.info(colored("Table originally with %i rows is truncated for top and bottom 5 only."%self.df_sort.shape[0], "cyan"))
+##        logger.info(colored("Consider filtering it with --n=x for the 1st x results or --filter-tags=foo using a value from your own EC2 tags.", "cyan"))
+#
+#    if self.analyzer.n!=-1:
+#      logger.info(colored("This table has been filtered for only the 1st %i underused results"%self.analyzer.n, "cyan"))
+#
+#    return context_all